Refactor and merge Reservations Controllers
