+++ maat-court-data-api/src/main/java/gov/uk/courtdata/reporder/controller/ReservationsController.java
@@ -2,27 +2,84 @@ package gov.uk.courtdata.reporder.controller;
 
 import gov.uk.courtdata.annotation.NotFoundApiResponse;
 import gov.uk.courtdata.annotation.StandardApiResponse;
+import gov.uk.courtdata.constants.CourtDataConstants;
 import gov.uk.courtdata.entity.ReservationsEntity;
+import gov.uk.courtdata.enums.LoggingData;
+import gov.uk.courtdata.model.authorization.AuthorizationResponse;
+import gov.uk.courtdata.prosecutionconcluded.helper.ReservationsRepositoryHelper;
 import gov.uk.courtdata.reporder.service.ReservationsService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.slf4j.MDC;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import uk.gov.justice.laa.crime.commons.common.Constants;
+
+import java.util.Optional;
+
+import static gov.uk.courtdata.enums.LoggingData.LAA_TRANSACTION_ID;
 
 @Slf4j
 @RestController
 @RequiredArgsConstructor
 @Tag(name = "Reservations", description = "Rest API for Reservations")
-@RequestMapping("/api/internal/v1/rep-orders/reservations")
+@RequestMapping("/api/internal/v1")
 public class ReservationsController {
     private final ReservationsService reservationsService;
 
-    @GetMapping(value = "/{id}")
+    private final ReservationsRepositoryHelper reservationsRepositoryHelper;
+
+    @GetMapping(value = "/assessment/reservations/{maatId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    @Operation(description = "Check if a given maatId is Locked through Reservations table")
+    @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
+    @NotFoundApiResponse
+    @StandardApiResponse
+    public ResponseEntity<Boolean> isMaatRecordLocked(
+            @PathVariable int maatId,
+            @Parameter(description = "Used for tracing calls")
+            @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
+        MDC.put(LAA_TRANSACTION_ID.getValue(), laaTransactionId);
+        log.info(String.format("Check if maatId is locked - %d {}", maatId));
+        return ResponseEntity.ok(reservationsRepositoryHelper.isMaatRecordLocked(maatId));
+    }
+
+    @GetMapping(value = "/assessment/reservations/recordname/{recordName}/recordid/{recordId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    @Operation(description = "Get the reservation status object of a record")
+    @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = AuthorizationResponse.class)))
+    @StandardApiResponse
+    public ResponseEntity<ReservationsEntity> getReservationByRecordNameAndRecordId(
+            @PathVariable String recordName, @PathVariable Integer recordId,
+            @Parameter(description = "Used for tracing calls")
+            @RequestHeader(value = Constants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
+        MDC.put(LoggingData.LAA_TRANSACTION_ID.getValue(), laaTransactionId);
+        log.info("Check reservation status - request received");
+        Optional<ReservationsEntity> reservationsEntity = reservationsRepositoryHelper.getReservationByRecordNameAndRecordId(recordName,recordId);
+        return reservationsEntity.map(c -> ResponseEntity.ok().body(c))
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @GetMapping(value = "/assessment/reservations/username/{username}", produces = MediaType.APPLICATION_JSON_VALUE)
+    @Operation(description = "Get the reservation status object of a record")
+    @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = AuthorizationResponse.class)))
+    @StandardApiResponse
+    public ResponseEntity<ReservationsEntity> getReservationByUsername(
+            @PathVariable String username,
+            @RequestHeader(value = Constants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
+        MDC.put(LoggingData.LAA_TRANSACTION_ID.getValue(), laaTransactionId);
+        log.info("Check reservation status - request received");
+        Optional<ReservationsEntity> reservationsEntity = reservationsRepositoryHelper.getReservationByUserName(username);
+        return reservationsEntity.map(c -> ResponseEntity.ok().body(c))
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @GetMapping(value = "/rep-orders/reservations/{id}")
     @Operation(description = "Retrieve a reservation")
     @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
     @StandardApiResponse
@@ -31,7 +88,7 @@ public class ReservationsController {
         return reservationsService.retrieve(id);
     }
 
-    @PostMapping
+    @PostMapping(value = "/rep-orders/reservations")
     @Operation(description = "Create a Reservation")
     @StandardApiResponse
     @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
@@ -40,7 +97,7 @@ public class ReservationsController {
         return ResponseEntity.ok().build();
     }
 
-    @PutMapping(value = "/{id}")
+    @PutMapping(value = "/rep-orders/reservations/{id}")
     @Operation(description = "Update a Reservation")
     @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
     @StandardApiResponse
@@ -50,7 +107,7 @@ public class ReservationsController {
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping(value = "/{id}")
+    @DeleteMapping(value = "/rep-orders/reservations/{id}")
     @Operation(description = "Delete a Reservation")
     @StandardApiResponse
     @ApiResponse(responseCode = "200", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
