+++ maat-court-data-api/src/main/resources/application.yaml
@@ -122,6 +122,7 @@ api-endpoints:
   assessments-domain: /api/internal/v1/assessment
   user-domain: /api/internal/v1/users
   application-domain: /api/internal/v1/application
+  reservations-domain: /api/internal/v1
 
 feature:
   postMvpEnabled: ${POST_MVP_ENABLED}
